Add files via upload
This is the first small block of code I wrote during my first day of learning how to use Python. This is an age generator where you provide your year of birth, and the code generates your age based on the year you were born in. Baby steps, but I would say pretty decent for the first day.