Loops and Lists
Day 3 saw me go on a freestyle session, exploring loops using the boolean functions "while," "for," and "in," as well as diving into how lists can be managed in Python. Here is also where the utilization of the # came into play, as I consistently took notes on what I was doing.